requirements.txt
requests==2.27.*
python-dotenv==0.19.*
python-telegram-bot==13.*